project/src/main/java/com/g28/model/Position.java
package com.g28.model;

public class Position 
{
    private int x;
    private int y;

    public Position(int x, int y) {
        this.x = x;
        this.y = y;
    }

    public int getX() { return x; }
    public int getY() { return y; }

    @Override
    public boolean equals(Object obj) {
        if (this == obj) return true;
        if (obj == null) return false;
        if (obj.getClass() != this.getClass()) return false;

        Position p = (Position) obj;
        return (x == p.getX() && y == p.getY());
    }
}